+++ servicio-transporte/src/main/java/com/backend/tpi_backend/servicio_transporte/repositories/TramoRepository.java
@@ -5,6 +5,6 @@ import java.util.List;
 import com.backend.tpi_backend.servicio_transporte.model.Ruta;
 import com.backend.tpi_backend.servicio_transporte.model.Tramo;
 
-public interface TramoRepository extends BaseRepository<Tramo, Long> {
+public interface TramoRepository extends BaseRepository<Tramo, Integer> {
     List<Tramo> findByRuta(Ruta ruta);
 }
