+++ servicio-transporte/src/main/java/com/backend/tpi_backend/servicio_transporte/repositories/TransportistaRepository.java
@@ -2,4 +2,4 @@ package com.backend.tpi_backend.servicio_transporte.repositories;
 
 import com.backend.tpi_backend.servicio_transporte.model.Transportista;
 
-public interface TransportistaRepository extends BaseRepository<Transportista, Long> { }
+public interface TransportistaRepository extends BaseRepository<Transportista, Integer> { }
